+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.24.0
 ### Features
 * new option repeatTimeoutKeys with default of HOLD+BACKSPACE that should have delay when deleting words
+* use setTimeout instead of alert to show exception that can be shown in terminal
+* allow to move cursor when regex formatter don't change length of the string
 ### Bugfix
 * fix delay when using arrow keys by filtering keys that have delay when hold
 * fix skipping lines that have emoji as last character inside formatting